feat(core) added azure messaging protocols
